create order Ok

+++ service/src/OnlineClothes.Domain/Entities/OrderItem.cs
@@ -7,12 +7,12 @@ public class OrderItem
 	public int OrderId { get; set; }
 	public string ProductSkuId { get; set; } = null!;
 	public int Quantity { get; set; }
-	public double Price { get; set; }
+	public decimal Price { get; set; }
 
 	[ForeignKey("OrderId")] public Order Order { get; set; } = null!;
 	[ForeignKey("ProductSkuId")] public ProductSku ProductSku { get; set; } = null!;
 
-	public double IntoMoney()
+	public decimal IntoMoney()
 	{
 		return Quantity * Price;
 	}
